ref(feedback): Refactor Feedback Bulk-Edit Comfirm dialog, and implement `bypass`
